# openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The record has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Invalid data provided
      tags:
        - Users
      security:
        - bearer: []
  /api/v1/users/individual:
    post:
      operationId: UsersController_createIndividual
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndividualUserDto'
      responses:
        '201':
          description: The record has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Invalid data provided
      tags:
        - Users
      security:
        - bearer: []
  /api/v1/users/{id}:
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: The record has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Invalid data provided
      tags:
        - Users
      security:
        - bearer: []
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Returns the user if exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Users
      security:
        - bearer: []
  /api/v1/users/{id}/activate:
    patch:
      operationId: UsersController_activate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateUserDto'
      responses:
        '200':
          description: The record has been successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Invalid data provided
      tags:
        - Users
      security:
        - bearer: []
  /api/v1/users/statistics:
    get:
      operationId: UsersController_statistics
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatistics'
      tags:
        - Users
      security:
        - bearer: []
  /api/v1/users/all:
    post:
      operationId: UsersController_findAll
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryUserDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageDto'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
      tags:
        - Users
      security:
        - bearer: []
...
  /api/v1/auth/email/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthEmailLoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseType'
        '401':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/v1/auth/email/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterLoginDto'
      responses:
        '201':
          description: User registered successfully
        '422':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/v1/auth/email/resend-verify-mail:
    post:
      operationId: AuthController_resendVerifyMail
      parameters: []
      responses:
        '200':
          description: Email sent successfully
        '422':
          description: Invalid State
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      security:
        - bearer: []
  /api/v1/auth/email/confirm:
    post:
      operationId: AuthController_confirmEmail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfirmEmailDto'
      responses:
        '204':
          description: User confirmed successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invalid user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/v1/auth/forgot/password:
    post:
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthForgotPasswordDto'
      responses:
        '204':
          description: Forgot request sent successfully
        '404':
          description: Invalid user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/v1/auth/reset/password:
    post:
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResetPasswordDto'
      responses:
        '204':
          description: Password reset successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invalid user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/v1/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      security:
        - bearer: []
    patch:
      operationId: AuthController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUpdateDto'
      responses:
        '200':
          description: User information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      security:
        - bearer: []
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponseType'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      security:
        - bearer: []
  /api/v1/auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '204':
          description: Logout successful
      tags:
        - Auth
      security:
        - bearer: []
  /api/v1/files/upload:
    post:
      operationId: FilesLocalController_uploadFile
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileType'
        '422':
          description: Missing file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Files
      security:
        - bearer: []
...
info:
  title: MOEPP CITES API
  description: MOEPP CITES API docs
  version: '1.0'
  contact: {}
...
servers:
  - url: https://moepp-cites.doozyx.com
    description: Testing server
  - url: http://localhost:8000
    description: Local server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          enum:
            - user
            - admin
            - approver
            - director
            - superAdmin
            - archiver
          type: string
        status:
          enum:
            - inactive
            - verified
            - pending
            - approved
            - rejected
            - expired
          type: string
        type:
          enum:
            - individual
            - company
          type: string
        language:
          enum:
            - en
            - mk
          type: string
      required:
        - email
        - password
        - role
        - status
        - type
        - language
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        countryId:
          type: number
        postalCode:
          type: string
      required:
        - street
        - city
        - state
        - countryId
        - postalCode
    Document:
      type: object
      properties:
        id:
          type: string
        issuingCountryId:
          type: number
        type:
          type: string
          enum:
            - personalId
            - passport
      required:
        - id
        - issuingCountryId
        - type
    Individual:
      type: object
      properties:
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
          nullable: true
        deletedAt:
          format: date-time
          type: string
          nullable: true
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        document:
          $ref: '#/components/schemas/Document'
        phoneNumber:
          type: string
      required:
        - createdAt
        - updatedAt
        - deletedAt
        - id
        - firstName
        - lastName
        - address
        - document
        - phoneNumber
    FileType:
      type: object
      properties:
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
          nullable: true
        deletedAt:
          format: date-time
          type: string
          nullable: true
        id:
          type: string
        name:
          type: string
      required:
        - createdAt
        - updatedAt
        - deletedAt
        - id
        - name
    Company:
      type: object
      properties:
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
          nullable: true
        deletedAt:
          format: date-time
          type: string
          nullable: true
        id:
          type: number
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        taxNumber:
          type: number
        identificationDocument:
          $ref: '#/components/schemas/FileType'
        identificationDocumentId:
          type: string
        contact:
          $ref: '#/components/schemas/Individual'
      required:
        - createdAt
        - updatedAt
        - deletedAt
        - id
        - name
        - address
        - taxNumber
        - identificationDocument
        - identificationDocumentId
        - contact
    User:
      type: object
      properties:
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
          nullable: true
        deletedAt:
          format: date-time
          type: string
          nullable: true
        id:
          type: number
        email:
          type: string
        role:
          type: string
          enum:
            - user
            - admin
            - approver
            - director
            - superAdmin
            - archiver
        status:
          type: string
          enum:
            - inactive
            - verified
            - pending
            - approved
            - rejected
            - expired
        statusReason:
          type: string
          nullable: true
        type:
          type: string
          enum:
            - individual
            - company
        individual:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Individual'
        company:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Company'
        language:
          type: string
          enum:
            - en
            - mk
      required:
        - createdAt
        - updatedAt
        - deletedAt
        - id
        - email
        - role
        - status
        - statusReason
        - type
        - individual
        - company
        - language
...
    CreateIndividualProfileDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        document:
          $ref: '#/components/schemas/DocumentDto'
        phoneNumber:
          type: string
      required:
        - firstName
        - lastName
        - address
        - document
        - phoneNumber
    CreateIndividualUserDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          enum:
            - user
            - admin
            - approver
            - director
            - superAdmin
            - archiver
          type: string
        status:
          enum:
            - inactive
            - verified
            - pending
            - approved
            - rejected
            - expired
          type: string
        individual:
          $ref: '#/components/schemas/CreateIndividualProfileDto'
        language:
          enum:
            - en
            - mk
          type: string
      required:
        - email
        - password
        - role
        - status
        - language
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          example: test1@doozyx.com
        password:
          type: string
          minLength: 6
        role:
          enum:
            - user
            - admin
            - approver
            - director
            - superAdmin
            - archiver
          type: string
        status:
          enum:
            - inactive
            - verified
            - pending
            - approved
            - rejected
            - expired
          type: string
        individual:
          $ref: '#/components/schemas/CreateIndividualProfileDto'
        language:
          enum:
            - en
            - mk
          type: string
        statusReason:
          type: string
    ActivateUserDto:
      type: object
      properties:
        status:
          enum:
            - inactive
            - verified
            - pending
            - approved
            - rejected
            - expired
          type: string
        statusReason:
          type: string
    UserStatistics:
      type: object
      properties:
        inactive:
          type: number
        verified:
          type: number
        pending:
          type: number
        approved:
          type: number
        rejected:
          type: number
        total:
          type: number
      required:
        - inactive
        - verified
        - pending
        - approved
        - rejected
        - total
    PageMetaDto:
      type: object
      properties:
        page:
          type: number
        limit:
          type: number
        itemCount:
          type: number
        pageCount:
          type: number
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
      required:
        - page
        - limit
        - itemCount
        - pageCount
        - hasPreviousPage
        - hasNextPage
    PageDto:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/PageMetaDto'
        data:
          type: array
          items:
            type: string
      required:
        - meta
        - data
    FilterUserDto:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
            enum:
              - user
              - admin
              - approver
              - director
              - superAdmin
              - archiver
        statuses:
          type: array
          items:
            type: string
            enum:
              - inactive
              - verified
              - pending
              - approved
              - rejected
              - expired
        email:
          type: string
        types:
          type: array
          items:
            type: string
            enum:
              - individual
              - company
    SortUserDto:
      type: object
      properties:
        orderBy:
          enum:
            - email
            - type
            - roleId
            - statusId
            - createdAt
            - updatedAt
            - deletedAt
          type: string
        order:
          default: ASC
          enum:
            - ASC
            - DESC
          type: string
      required:
        - orderBy
    QueryUserDto:
      type: object
      properties:
        page:
          type: number
          default: 1
          minimum: 1
        limit:
          type: number
          default: 10
          minimum: 1
          maximum: 50
        filter:
          $ref: '#/components/schemas/FilterUserDto'
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortUserDto'
...
    AuthEmailLoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@doozyx.com
        password:
          type: string
      required:
        - email
        - password
    LoginResponseType:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        tokenExpires:
          type: number
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - refreshToken
        - tokenExpires
        - user
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        statusCode:
          type: number
      required:
        - message
        - error
        - statusCode
    AuthRegisterLoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@doozyx.com
        password:
          type: string
          minLength: 6
        type:
          enum:
            - individual
            - company
          type: string
        language:
          enum:
            - en
            - mk
          type: string
      required:
        - email
        - password
        - type
        - language
    AuthConfirmEmailDto:
      type: object
      properties:
        hash:
          type: string
      required:
        - hash
    AuthForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    AuthResetPasswordDto:
      type: object
      properties:
        password:
          type: string
        hash:
          type: string
      required:
        - password
        - hash
    RefreshResponseType:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        tokenExpires:
          type: number
      required:
        - token
        - refreshToken
        - tokenExpires
    AuthUpdateDto:
      type: object
      properties:
        password:
          type: string
          minLength: 6
        oldPassword:
          type: string
...